backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pypdf